lutris: fix RPCS3 AppImage
A lot of runners are AppImages and these due to their fuse + FHS dependencies, won't work as expected in Nix. This commit implements a workaround that Lutris also employes for flatpaks where they unpack the image and then they run it. This is done via an environment variable passed to the Lutris executable.

This commit also brings all dependencies needed for RPCS3 emulator to work as expected.